Clean up for more consistent modules
- Rename module and procedures to be more consistnet with coding standard
- Debug printout adjustment
