+++ project/ANT.h
@@ -1,0 +1,26 @@
+/******************************************************************************
+File: ANT.h
+
+This header file contains the macros and function declarations for ANT communication
+
+******************************************************************************/
+
+
+#ifndef ANT_H__
+#define ANT_H__
+
+/*******************************************************************************
+															GENERAL INCLUDES
+*******************************************************************************/
+
+#include <stdint.h>
+#include "bsp.h"
+
+
+/*******************************************************************************
+															   PROCEDURES
+*******************************************************************************/
+
+void ANT_init(void);
+
+#endif  // ANT_H__
