+++ project/LIS2DH12.c
@@ -66,6 +66,10 @@ typedef struct {
 *******************************************************************************/
 lis2dh12_instance_t accel_inst = {0}; // Instance of the LIS2DH12
 
+/*NOTE: Uncomment these lines for more debug info*/
+#define SPI_DEBUG_INFO
+#define ACCEL_DEBUG_INFO
+
 #define SPI_INSTANCE  0 //SPI instance index
 const nrf_drv_spi_t spi = NRF_DRV_SPI_INSTANCE(SPI_INSTANCE); //SPI instance
 volatile bool spi_xfer_done; //Flag used to indicate that SPI instance completed the transfer
@@ -546,7 +550,7 @@ bool ACCEL_init(void)
 	accel_write_block(&block_command);
 
 	#ifdef ACCEL_DEBUG_INFO
-	NRF_LOG_INFO("Accelerometer Initializing...");
+	NRF_LOG_INFO("Verifying...");
 	#endif
 
 	control_block_t config_block_r = {0};
